Разделил файл utils.js на несколько и вынес в отдельную папку
