+++ src/utils/common.js
@@ -1,0 +1,12 @@
+
+//Использование замыканий для увеличения числа на единицу
+const increaseNumber = () => {
+  let num = 1;
+  return () => num++;
+};
+
+function getRandomArrayElement(items){
+  return items[Math.floor(Math.random() * items.length)];
+}
+
+export { increaseNumber, getRandomArrayElement};
