+++ src/utils/date.js
@@ -33,14 +33,4 @@ function isEventExpired(dueDate){
   return dueDate && dayjs().isAfter(dueDate, 'D');
 }
 
-//Использование замыканий для увеличения числа на единицу
-const increaseNumber = () => {
-  let num = 1;
-  return () => num++;
-};
-
-function getRandomArrayElement(items){
-  return items[Math.floor(Math.random() * items.length)];
-}
-
-export { increaseNumber, getRandomArrayElement, humanizeDate, isEventExpired, getTimeDifference};
+export { humanizeDate, isEventExpired, getTimeDifference };
